refactor: move logic to helper functions
